requirements.txt
git+https://github.com/ZippyMatch/django-datetime-widget.git
alabaster==0.7.9
amqp==1.4.9
anyjson==0.3.3
argh==0.26.2
async-notifications==0.0.3
Babel==2.3.4
backports-abc==0.4
billiard==3.3.0.23
cairocffi==0.7.2
CairoSVG==1.0.22
celery==3.1.23
cffi==1.8.3
cssselect==0.9.2
Django==1.10
django-ckeditor==5.1.1
django-password-reset==0.9
django-registration==2.1.2
django-reversion==2.0.6
djangoajax==2.3.6
docutils==0.12
html5lib==0.999999999
imagesize==0.7.1
Jinja2==2.8
kombu==3.0.35
livereload==2.4.1
lxml==3.6.4
MarkupSafe==0.23
numpy==1.11.2
pathtools==0.1.2
port-for==0.3.1
psycopg2==2.6.2
pycparser==2.14
Pygments==2.1.3
Pyphen==0.9.4
pytz==2016.6.1
PyYAML==3.12
savReaderWriter==3.4.2
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.4.6
sphinx-autobuild==0.6.0
sphinx-rtd-theme==0.1.9
tinycss==0.3
tornado==4.4.1
watchdog==0.8.3
WeasyPrint==0.31
webencodings==0.5
django-bootstrap-form==3.4
django-debug-toolbar==1.8
